feat: 前置初始化数据阻塞
