+++ src/hooks/useEffectLoading.ts
@@ -1,16 +1,24 @@
-import { useEffect, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 
 type RequestFn<T> = (param?: T) => Promise<void>
 
 export function useEffectLoading <T> (deps: unknown[], requestFn: RequestFn<T>, param?: T) {
-  const [loading, setLoading] = useState(false)
+  const [loading, setLoading] = useState(true)
+
+  const handleRequest = useCallback(async () => {
+    setLoading(true)
+    try {
+      await requestFn(param)
+    } catch (error) {
+      console.log(error)
+    } finally {
+      setLoading(false)
+    }
+  }, deps)
 
   useEffect(() => {
-    setLoading(true)
-    void requestFn(param)
-      .catch((error) => { console.log(error) })
-      .finally(() => { setLoading(false) })
-  }, deps)
+    void handleRequest()
+  }, [handleRequest])
 
   return { loading }
 }
