+++ src/main.tsx
@@ -1,18 +1,13 @@
 import { StrictMode } from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
-import DisableHotKeys from './components/disableHotKeys'
-import PlayerInitiator from './PlayerInitiator'
+import './core/player'
 import './style/index.less'
 import './main.css'
 
 const root: HTMLElement = document.getElementById('root')!
 ReactDOM.createRoot(root).render(
   <StrictMode>
-    <DisableHotKeys>
-      <PlayerInitiator>
-        <App />
-      </PlayerInitiator>
-    </DisableHotKeys>
+    <App />
   </StrictMode>
 )
