+++ src/store/persist.ts
@@ -1,17 +1,14 @@
 import { atom } from 'jotai'
-import { selectDB, updateDB } from '@/utils'
+import { updateDB } from '@/utils'
 
 interface PersistParam<T> {
   cacheName: string
-  enableInitialCache: boolean
   initialValue: T
   debounceMS?: number
   throttleMS?: number
 }
 
-export const createAtomWithIndexedDB = <T>({ enableInitialCache, cacheName, initialValue, debounceMS, throttleMS }: PersistParam<T>) => {
-  let lock = true
-
+export const createAtomWithIndexedDB = <T>({ cacheName, initialValue, debounceMS, throttleMS }: PersistParam<T>) => {
   let timer = 0
 
   const baseAtom = atom<T>(initialValue)
@@ -19,8 +16,6 @@ export const createAtomWithIndexedDB = <T>({ enableInitialCache, cacheName, init
   const cacheAtom = atom(
     (get) => get(baseAtom),
     (_, set, value: T) => {
-      if (lock) return
-
       set(baseAtom, value)
 
       if (debounceMS != null && debounceMS > 0) {
@@ -39,23 +34,6 @@ export const createAtomWithIndexedDB = <T>({ enableInitialCache, cacheName, init
       }
     }
   )
-
-  baseAtom.onMount = (setAtom) => {
-    if (!enableInitialCache) {
-      lock = false
-      return
-    }
-
-    void selectDB<T>(cacheName)
-      .then((value) => {
-        if (value != null) {
-          setAtom(value)
-        }
-      })
-      .finally(() => {
-        lock = false
-      })
-  }
 
   return cacheAtom
 }
