+++ src/store/playerController.ts
@@ -5,7 +5,6 @@ import { PlayType } from '@/core/player'
 export const PLAY_TYPE_CACHE_NAME = 'playType'
 
 export const playTypeAtom = createAtomWithIndexedDB<PlayType>({
-  enableInitialCache: true,
   cacheName: PLAY_TYPE_CACHE_NAME,
   initialValue: PlayType.loop,
   debounceMS: 250
@@ -15,7 +14,6 @@ export const MUTE_CACHE_NAME = 'mute'
 
 export const muteAtom = createAtomWithIndexedDB<boolean>({
   cacheName: MUTE_CACHE_NAME,
-  enableInitialCache: true,
   initialValue: false,
   debounceMS: 250
 })
@@ -24,7 +22,6 @@ export const VOLUME_CACHE_NAME = 'volume'
 
 export const volumeAtom = createAtomWithIndexedDB<number>({
   cacheName: VOLUME_CACHE_NAME,
-  enableInitialCache: true,
   initialValue: 60,
   debounceMS: 250
 })
@@ -33,7 +30,6 @@ export const PROGRESS_CACHE_NAME = 'progress'
 
 export const progressAtom = createAtomWithIndexedDB<number>({
   cacheName: PROGRESS_CACHE_NAME,
-  enableInitialCache: true,
   initialValue: 0,
   throttleMS: 1000
 })
