+++ src/views/discovery/recommend/index.tsx
@@ -1,7 +1,8 @@
 import { memo, useCallback, useEffect, useMemo, useRef } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
 import { useAtomValue } from 'jotai'
-import { useRecommend, accountAtom, useFetchPlaylists } from '@/store'
+import { accountAtom } from '@/store'
+import { useFetchPlaylists, useRecommend } from '@/hooks'
 import Card from './card'
 import SwiperComponent from '@/components/swiper'
 import { CUSTOM_IMG } from '@/utils'
