+++ src/views/musicDetail/kinds/customDetail.tsx
@@ -1,7 +1,7 @@
 import { memo, useCallback, useEffect } from 'react'
 import { useAtomValue } from 'jotai'
-import { playlistInfoAtom, useFetchPlaylists } from '@/store'
-import { useMusicDetail } from '@/hooks'
+import { playlistInfoAtom } from '@/store'
+import { useFetchPlaylists, useMusicDetail } from '@/hooks'
 import MusicDetailHeader from '../components/header'
 import MusicDetailTab from '../components/tab'
 import { CUSTOM_ID, CUSTOM_IMG, CUSTOM_NAME } from '@/utils'
