+++ src/views/musicDetail/kinds/musicDetail.tsx
@@ -1,7 +1,7 @@
 import { memo, useCallback } from 'react'
 import { useAtomValue } from 'jotai'
-import { playlistInfoAtom, useFetchPlaylists } from '@/store'
-import { useMusicDetail, useEffectLoading } from '@/hooks'
+import { playlistInfoAtom } from '@/store'
+import { useMusicDetail, useEffectLoading, useFetchPlaylists } from '@/hooks'
 import MusicDetailHeader from '../components/header'
 import MusicDetailTab from '../components/tab'
 import { getPlaylistDetail } from '@/api'
